requirements.txt
# Base dependencies
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2

# Web
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# AI/ML - Install PyTorch separately
transformers>=4.36.0
huggingface-hub==0.20.3
pyautogen==0.7.5
